Add SASS and update extensions
