+++ pages/index.jsx
@@ -1,0 +1,41 @@
+import Head from "next/head"
+import Image from "next/image"
+import Card from "components/Card"
+import styles from "styles/Home.module.scss"
+
+export default function Home() {
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>Climatempo - Code Challenge - Vaga FrontEnd</title>
+        <meta
+          name="description"
+          content="Code Challenge Vaga Front da Climatempo"
+        />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <main className={styles.main}>
+        <div className={styles.grid}>
+          <Card></Card>
+          <Card></Card>
+          <Card></Card>
+          <Card></Card>
+        </div>
+      </main>
+
+      <footer className={styles.footer}>
+        <a
+          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Powered by{" "}
+          <span className={styles.logo}>
+            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
+          </span>
+        </a>
+      </footer>
+    </div>
+  )
+}
