+++ styles/Home.module.scss
@@ -1,24 +1,26 @@
+%flex-center {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
 .container {
   padding: 0 2rem;
 }
 
 .main {
+  @extend %flex-center;
   min-height: 100vh;
   padding: 4rem 0;
   flex: 1;
-  display: flex;
   flex-direction: column;
-  justify-content: center;
-  align-items: center;
 }
 
 .footer {
-  display: flex;
+  @extend %flex-center;
   flex: 1;
   padding: 2rem 0;
   border-top: 1px solid #eaeaea;
-  justify-content: center;
-  align-items: center;
 }
 
 .footer a {
